server: clean up coyote-core; move shutdown and instance id there; fix Cargo.toml
