+++ Cargo.toml
@@ -30,7 +30,7 @@ axum-extra.workspace = true
 byteorder.workspace = true
 constant_time_eq.workspace = true
 coyote-cache.workspace = true
-coyote-core = { version = "0.0.1", path = "crates/coyote-core" }
+coyote-core.workspace = true
 coyote-derive.workspace = true
 coyote-error.workspace = true
 coyote-idempotency.workspace = true
@@ -66,7 +66,7 @@ schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
-sysinfo = "0.38.2"
+sysinfo.workspace = true
 tap.workspace = true
 tokio.workspace = true
 tokio-util.workspace = true
@@ -133,6 +133,7 @@ constant_time_eq = "0.4"
 coyote.path = "."
 coyote-cache.path = "crates/cache"
 coyote-client.path = "clients/rust"
+coyote-core.path = "crates/coyote-core"
 coyote-derive.path = "crates/coyote-derive"
 coyote-error.path = "crates/coyote-error"
 coyote-idempotency.path = "crates/idempotency"
@@ -193,6 +194,7 @@ serde_json = { version = "1.0.74", features = ["preserve_order", "raw_value"] }
 serde_path_to_error = "0.1.7"
 sha2 = "0.10.2"
 syn = "2.0.48"
+sysinfo = "0.38.2"
 tap = "1"
 tempfile = "3.24.0"
 test-utils.path = "crates/test-utils"
