+++ server/src/otel.rs
@@ -1,9 +1,7 @@
 use std::time::Duration;
 
-use coyote::{
-    INSTANCE_ID,
-    cfg::{self, ConfigurationInner},
-};
+use coyote::cfg::{self, ConfigurationInner};
+use coyote_core::INSTANCE_ID;
 use opentelemetry::{InstrumentationScope, trace::TracerProvider as _};
 use opentelemetry_otlp::WithExportConfig;
 use opentelemetry_sdk::{
